EarningsCallReader_L1Summarization_11thNov/requirements.txt
mistralai
openai
streamlit
streamlit-pdf-viewer
pyyaml
pandas
langchain
langchain_text_splitters
langchain_experimental
langchain_openai
langchain_core
langchain_community
auth 
openpyxl
xlsxwriter
nest_asyncio